--- qa/smoke-test-shaded/src/test/java/org/elasticsearch/shaded/test/ShadedTest.java
@@ -1,49 +1,0 @@
-/*
- * Licensed to Elasticsearch under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.elasticsearch.shaded.test;
-
-import org.apache.lucene.util.LuceneTestCase;
-import org.junit.Test;
-
-/**
- */
-public class ShadedTest extends LuceneTestCase {
-
-    @Test
-    public void testLoadShadedClasses() throws ClassNotFoundException {
-        Class.forName("org.elasticsearch.common.collect.ImmutableList");
-        Class.forName("org.elasticsearch.common.joda.time.DateTime");
-        Class.forName("org.elasticsearch.common.util.concurrent.jsr166e.LongAdder");
-    }
-
-    @Test(expected = ClassNotFoundException.class)
-    public void testGuavaIsNotOnTheCP() throws ClassNotFoundException {
-        Class.forName("com.google.common.collect.ImmutableList");
-    }
-
-    @Test(expected = ClassNotFoundException.class)
-    public void testJodaIsNotOnTheCP() throws ClassNotFoundException {
-        Class.forName("org.joda.time.DateTime");
-    }
-
-    @Test(expected = ClassNotFoundException.class)
-    public void testjsr166eIsNotOnTheCP() throws ClassNotFoundException {
-        Class.forName("com.twitter.jsr166e.LongAdder");
-    }
-}
